requirements.txt
Flask==3.1.2
gunicorn
psycopg2-binary==2.9.11
bcrypt==5.0.0
Flask-Bcrypt==1.0.1
requests==2.32.5
user-agents



